requirements.txt
cachetools==6.2.1
click==8.3.0
cloudpickle==3.1.1
cubinlinker-cu11==0.3.0.post3
cuda-bindings==11.8.7
cuda-python==11.8.7
cudf-cu11==25.6.0
cuml-cu11==25.6.0
cupy-cuda11x==13.6.0
cuvs-cu11==25.6.1
dask==2025.5.0
dask-cuda==25.6.0
dask-cudf-cu11==25.6.0
distributed==2025.5.0
distributed-ucxx-cu11==0.44.0
fastrlock==0.8.3
fsspec==2025.10.0
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
libcudf-cu11==25.6.0
libcuml-cu11==25.6.0
libcuvs-cu11==25.6.1
libkvikio-cu11==25.6.0
libraft-cu11==25.6.0
librmm-cu11==25.6.0
libucx-cu11==1.18.1
libucxx-cu11==0.44.0
llvmlite==0.44.0
locket==1.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numba==0.61.2
numba-cuda==0.11.0
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.575.51
nvtx==0.2.13
packaging==25.0
pandas==2.2.3
partd==1.4.2
psutil==7.1.3
ptxcompiler-cu11==0.8.1.post3
pyarrow==19.0.1
Pygments==2.19.2
pylibcudf-cu11==25.6.0
pylibraft-cu11==25.6.0
pynvml==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
raft-dask-cu11==25.6.0
rapids-dask-dependency==25.6.0
rapids-logger==0.1.19
rich==14.2.0
rmm-cu11==25.6.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sortedcontainers==2.4.0
tblib==3.2.1
threadpoolctl==3.6.0
toolz==1.1.0
tornado==6.5.2
treelite==4.4.1
typing_extensions==4.15.0
tzdata==2025.2
ucx-py-cu11==0.44.0
ucxx-cu11==0.44.0
urllib3==2.5.0
zict==3.0.0
zipp==3.23.0
